feat(sync): take server creation out of the app, and give paint sync a home
Nobody should be creating servers from the app right now. The Servers tab is
where that happened, so it goes, and with it the experimental toggle whose only
job was revealing it.

Paint sync was living on that page, which is no longer where it belongs: it is
finished, on by default, and for everyone. It moves to Settings → Paint sync,
beside the switch that turns it off.

- `Servers.tsx` deleted: the registry rows, the cloud fleet, provisioning and
  Create server all go with it. The backend commands stay, and the control
  plane still refuses registration and provisioning to anyone without an
  invite, so this closes a door that was already locked.
- The sync panel is lifted into `Settings/PaintSync.tsx`, minus the invite-code
  form. There is nothing left for a code to unlock, and sync claims its own
  account — so where the form was, the panel now says it starts on its own.
- Join-by-address is no longer hidden. Joining a server is not hosting one.
- `experimental` is gone from the config, the API and the tour: it gated the
  tab, that button and itself, and nothing else.
