+++ src/i18n/locales/es.ts
@@ -1092,6 +1092,9 @@ export const es: Translation = {
   "sync.lastPulled": "Última comprobación {{ago}}. Se repite sola cuando pulsas Jugar.",
   "sync.neverPulled": "Aún no has descargado las pinturas de los demás",
   "sync.neverPulledWhy": "Hasta que lo hagas, los otros pilotos aparecen con motos por defecto aunque hayan publicado las suyas.",
+  "sync.notStartedTitle": "Aún no ha empezado",
+  "sync.notStartedWhy":
+    "La sincronización de pinturas se configura sola la primera vez que ruedas en línea — no hay nada que rellenar. Abre MX Bikes, únete a un servidor, y esto se completará solo.",
   "sync.oversized_one": "{{count}} pintura es demasiado grande para compartirla, así que los demás pilotos no la verán.",
   "sync.oversized_other": "{{count}} pinturas son demasiado grandes para compartirlas, así que los demás pilotos no las verán.",
   "sync.skippedBikes_one": "{{count}} moto no se publicó — tienes más de las que podemos guardar.",
