+++ src/i18n/locales/pt-BR.ts
@@ -1092,6 +1092,9 @@ export const ptBR: Translation = {
   "sync.lastPulled": "Última verificação {{ago}}. Ela roda sozinha quando você aperta Jogar.",
   "sync.neverPulled": "Você ainda não baixou as pinturas dos outros",
   "sync.neverPulledWhy": "Até fazer isso, os outros pilotos aparecem com motos padrão mesmo tendo publicado as deles.",
+  "sync.notStartedTitle": "Ainda não começou",
+  "sync.notStartedWhy":
+    "A sincronização de pinturas se configura sozinha na primeira vez que você anda online — não há nada para preencher. Abra o MX Bikes, entre em um servidor, e o resto aparece aqui sozinho.",
   "sync.oversized_one": "{{count}} pintura é grande demais para compartilhar, então os outros pilotos não vão vê-la.",
   "sync.oversized_other": "{{count}} pinturas são grandes demais para compartilhar, então os outros pilotos não vão vê-las.",
   "sync.skippedBikes_one": "{{count}} moto não foi publicada — você tem mais do que conseguimos guardar.",
